move `is_allowed_content_path` and `resolve_allowed_paths` to `scanner` folder
