+++ crates/oxide/src/scanner/fixtures/binary-extensions.txt
