+++ crates/oxide/src/scanner/fixtures/ignored-extensions.txt
