+++ crates/oxide/src/scanner/fixtures/ignored-files.txt
