+++ crates/oxide/src/scanner/fixtures/template-extensions.txt
