+++ crates/oxide/src/scanner/mod.rs
@@ -1,0 +1,1 @@
+pub mod allowed_paths;
